Enables Android Oreo support with a workaround
Adnroid Oreo adds a restriction on registering Intent Receivers
statically within AndroidManifest.xml
This adds a workaround to register the app as a Connected CarAPI app.
When the Connected app connects to the car, it is supposed to send a
unique Intent that might make it through the background restriction.
But it also opens up the CarAPI Content Provider, which is now taken
as a signal to start probing for the car connection.