Fixes #15 - Handles multiple notifications at the same time better
Apps with grouped notifications are observed to send the entire group every time a new notification shows up
However, if the new notification is at the end of the group, a previous notification within the group will updateNotificationList()
and this will pull the entire group into the current list of notifications,
which blocks the new notification at the end from popping up.
Now, there's a discrete list of popups that have been shown to fix that.