Only updates Global Metadata for specific multimedia sources
Listens to the car's Multimedia Source to guess if the car is listening
to the phone's audio. However, the IDs seem to be different than the
CarDataConverter IDs for the sources, and perhaps Mini has a different
set of IDs.
For example, the car's Aux port is coming from ID 21, which is
documented to be AS_TEL_Ring, while the Bluetooth is coming from ID 32,
which is documented to be AS_ES_USB_3.

Without this change in place, a phone that has the app running but is
not the selected media source in the car will overwrite the car's
media source statusbar label whenever the phone starts playing media.
For example, a passenger watching a video on Reddit while the driver
is listening to the radio instead of the car app.