Fixes #41 by synchronizing on the NotificationsState during updates
WhatsApp sends a GroupSummary notification, which triggers an full
refresh of NotificationState.notifications. This is very close after
the direct message Notification triggers a message to the car to look
up a notification by key to show as a popup. When the car thread tries
to look up the notification, it is in the middle of refreshing, and thus
empty until it adds all of the notifications back at once.

This forces the car thread to wait until the refresh is finished.