Defers creating the avContext until InstanceID is known
With A2DPListener starting MainService for a Bluetooth connection
and with IDriveConnectionListener watching for BCL Reports
the InstanceID is unlikely to stay null for long.
This disables creating the AvContext if InstanceID is unknown,
because the null-coalesced InstanceID is wrong and will be corrected
when the next BCL Report happens.