Resets the infinitely-spinning AM app icons
The AM app icons are used to trigger actions while having
customizable icons and names. The car expects a full
RHMI app to replace them when clicked, and shows a
spinning loading icon until that happens. Since that
never happens, it spins forever.
It turns out that calling am_registerApp with
the same parameters replaces the spinning icon
with a new icon