Sorts out radio apps to the Radio section
Makes sure to not do it in ID4: ID4's Radio menu switches
playback to the last Radio source, such as FM, while this app
is a general Multimedia app. A user trying to go into the Radio
menu to switch app control would switch the car from Bluetooth to.. FM
music.