Cleans Bidi Isolate characters from car strings
Like emoji, the car doesn't seem to support bidi isolate markers.
This mainly relates to a contact's username/number (which is wrapped in
<FSI>+1-234-456<PDI>) causing the surrounding text to default to RTL.
Latin script uses the weak LTR and looks correct, but the punctuation
is neutral and winds up in RTL order.
Fixes (at least) one of the problems in #140