Only skips to a single folder in short lists
A user encountered a strange edgecase in single folder handling:
Single Folder handling was built to ignore playable entries before the
single folder, because of DSub's Play All entries. This user had a
Spotify Library with many playable playlist entries and a single
playlist folder at the end of the list, which was triggering this
Single Folder handling.
This change caps the maximum size of a list to be considered for
Single Folder handling.