Skips avcontext creation over USB
Audio Context is generally not supported over USB, and
the app doesn't request that the car listens to context over USB.
However, if the app was previously running over BT and had context,
the car would remember and automatically grant audio context over USB
when the app created the context.
This change doesn't even create the context if it isn't going to work,
to prevent the car from remembering and trying to grant context.