Restarts the CarCapabilities thread if it is dead during startup
During connection startup, the Apps tunnel sometimes starts and dies and restarts
And if CarCapabilities was running previously, it'll die when the connection goes down
But if it hasn't finished loading the car's language, which requires a CDS callback,
the CarCapabilities previously wouldn't be restarted to finish loading the language
and then the rest of the apps will never get started.
The CDS callback might encounter difficulties when the Apps protocol has split-brained,
because the async response won't come back to the TCP socket properly.