Fixes occasional app recreate after notification popup
Because ID5PopupView is technically a separate RHMI app, the
PhoneNotifications doesn't notice that it's showing and so
it doesn't reset the autorecreate timer. Both the popup autoclose
and the notification autorecreate timer would expire at 10s, and so
sometimes the notifications app would recreate while the popup
is closing to show the notifications menu.