Calmer representation of some values (#847)
Due to the appearance and disappearance of some characters, the display in the car is unstable. This particularly affects the steering and acceleration values. Regarding the minus sign and the arrows. This PR reduces this behavior.
In addition, the links to the nightly builds in the `readme` are presented more clearly.